Moving the `find_terminal_view` from `visualmetricsWrapper` to `terminalHelper`
